pyproject.toml
[tool.poetry]
name = "examplepython"
version = "0.1.0"
description = ""
authors = ["nester256", "nestvlad2004@gmail.com"]

#packages = [
#    { include = "webapp" }
#]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = ">=0.34.0,<0.35.0"
pydantic = ">=2.10.5,<3.0.0"
pydantic-settings = ">=2.7.1,<3.0.0"
orjson = ">=3.10.15,<4.0.0"
sqlalchemy = "2.0.23"
python-multipart = "0.0.19"
starlette-prometheus = "0.9.0"
starlette-context = "0.3.6"
PyYAML = "6.0.1"
fastapi = "^0.115.6"



[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
isort = "^5.13.2"
mypy = "^1.14.1"

